requirements.txt
# Core platform dependencies
pydantic>=2.0
pyyaml>=6.0

# Development / test
pytest>=7.0
pytest-cov>=4.0
mypy>=1.0
